chore: rename from template lol
