+++ dojo/base/miner.py
@@ -5,7 +5,7 @@ import traceback
 from bittensor.btlogging import logging as logger
 
 from commons.utils import serve_axon
-from template.base.neuron import BaseNeuron
+from dojo.base.neuron import BaseNeuron
 
 
 class BaseMinerNeuron(BaseNeuron):
