+++ dojo_cli.py
@@ -8,8 +8,8 @@ from prompt_toolkit import PromptSession, prompt
 from prompt_toolkit.completion import FuzzyCompleter, WordCompleter
 from rich.console import Console
 
-from template import get_dojo_api_base_url
-from template.utils.config import source_dotenv
+from dojo import get_dojo_api_base_url
+from dojo.utils.config import source_dotenv
 
 DOJO_API_BASE_URL = get_dojo_api_base_url()
