+++ tests/unit_testing/test_scoring.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 import torch
 
-from template.protocol import FeedbackRequest, RankingCriteria, TaskType
+from dojo.protocol import FeedbackRequest, RankingCriteria, TaskType
 
 # # Remove the default loguru handler
 # logger.remove()
@@ -26,7 +26,7 @@ def scoring_module():
     # ensure we import them depending on mock_env_var so the ValueError doesn't
     # get raised
     from commons.scoring import Scoring
-    from template.protocol import (
+    from dojo.protocol import (
         CodeAnswer,
         CompletionResponses,
         FeedbackRequest,
@@ -53,7 +53,7 @@ def mock_response(
     cid: str = "",
     rank_id: int = 0,
 ):
-    from template.protocol import CodeAnswer, CompletionResponses, FileObject
+    from dojo.protocol import CodeAnswer, CompletionResponses, FileObject
 
     return CompletionResponses(
         model=model,
@@ -67,7 +67,7 @@ def mock_response(
 
 
 def mock_request(hotkey: str | None = None, scores: list[float] | None = None):
-    from template.protocol import MultiScoreCriteria
+    from dojo.protocol import MultiScoreCriteria
 
     axon = bt.TerminalInfo(hotkey=hotkey)
     prompt = "Write a hello world program in python"
@@ -293,7 +293,7 @@ def mock_request_spm(
     """
     Dynamically generates miner responses using separate rank_ids and cids.
     """
-    from template.protocol import FeedbackRequest, TaskType
+    from dojo.protocol import FeedbackRequest, TaskType
 
     axon = bt.TerminalInfo(hotkey=hotkey)
     prompt = "Write a hello world program in python"
